requirements.txt
faster-whisper
ffmpeg-python
pytubefix
tk